pyproject.toml
[project]
name = "i-memory"
version = "1.0.0"
description = ""
authors = [
    {name = "eric",email = "whatisjava@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "psycopg2 (>=2.9.11,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "colorlog (>=6.10.1,<7.0.0)",
    "pyrootutils (>=1.0.4,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "injector (>=0.24.0,<0.25.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "fastapi (>=0.128.0,<0.129.0)",
    "pypdf (>=6.6.2,<7.0.0)",
    "mammoth (>=1.11.0,<2.0.0)",
    "openai (>=2.16.0,<3.0.0)",
    "markdownify (>=1.2.2,<2.0.0)",
    "numpy (>=2.4.1,<3.0.0)",
    "boto3 (>=1.42.37,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "redis (>=7.1.0,<8.0.0)",
    "jieba (>=0.42.1,<0.43.0)",
    "regex (>=2026.1.15,<2027.0.0)",
    "langchain (>=1.2.8,<2.0.0)",
    "langchain-openai (>=1.1.7,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
